requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
pymongo==4.7.3
bcrypt==4.1.3
python-dotenv==1.0.1
pytest==8.2.2
